requirements.txt
# FastAPI & Server
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-multipart==0.0.12

# AI Integration
google-generativeai==0.8.3

# Data Validation
pydantic==2.10.0
pydantic-settings==2.6.0

# HTTP Client for async requests
httpx==0.27.2

# Rate Limiting
slowapi==0.1.9

# Logging
python-json-logger==3.2.0

# Testing
pytest==8.3.0
pytest-asyncio==0.24.0
pytest-cov==6.0.0

# Code Quality (optional, for development)
black==24.10.0
ruff==0.8.0
